# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "gateway-server",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.3.0"
  },
  "paths": {
...
    "/admin/audit-events": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/billing/outbox": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/billing/outbox/{event_id}/retry": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/billing/webhooks": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
...
    "/admin/model-prices": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/model-prices/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "patch": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/model-routes": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/model-routes/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "patch": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/policies": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/policies/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "patch": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/projects": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/projects/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "patch": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/providers": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/providers/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "patch": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/providers/{id}/check": {
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/quota-limits": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/quota-limits/{id}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "patch": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
...
    "/admin/summary": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/system": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/tenants": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/usage/events": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/usage/mcp-invocations": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/usage/reservations": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/usage/series": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/usage/summary": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
    "/admin/virtual-keys": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
...
    "/auth/tenants/{tenant_id}/members": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not authorized"
          }
        }
      }
    },
...
  },
...
}